Algorithm logic (#131)
- Algorithm logic to run them in parallel.
- Introducing sync algorithms
- Introducing async algorithms
- Adding all luna algorithms as async algorithm

+++ luna_bench/components/features/optsol_feature.py
@@ -78,7 +78,7 @@ class OptSolFeature(IFeature):
     ...     print(f"Optimal value: {result.best_sol}")
     """
 
-    max_runtime: int | None = None  # define max runtime in seconds
+    max_runtime: float | None = None  # define max runtime in seconds
 
     def run(self, model: Model) -> OptSolFeatureResult:
         """
@@ -111,7 +111,7 @@ class OptSolFeature(IFeature):
         """
         scip_model = ScipModel()
         if self.max_runtime is not None:
-            scip_model.setParam("limits/time", self.max_runtime)
+            scip_model.setParam("limits/time", self.max_runtime)  # type: ignore[no-untyped-call]
 
         with tempfile.NamedTemporaryFile(suffix=".lp", delete=False) as tmp:
             path = Path(tmp.name)
@@ -121,18 +121,18 @@ class OptSolFeature(IFeature):
                 model,
                 filepath=path,
             )
-            scip_model.readProblem(path)
+            scip_model.readProblem(path)  # type: ignore[no-untyped-call]
         finally:
             if path.exists():
                 path.unlink()
 
-        scip_model.optimize()
-        if scip_model.getStatus() == "infeasible":
+        scip_model.optimize()  # type: ignore[no-untyped-call]
+        if scip_model.getStatus() == "infeasible":  # type: ignore[no-untyped-call]
             raise InfeasibleModelError
 
         # translate model to
         pre_terminated = False
-        if scip_model.getStatus() == "timelimit":
+        if scip_model.getStatus() == "timelimit":  # type: ignore[no-untyped-call]
             pre_terminated = True
 
-        return OptSolFeatureResult(best_sol=scip_model.getObjVal(), pre_terminated=pre_terminated)
+        return OptSolFeatureResult(best_sol=scip_model.getObjVal(), pre_terminated=pre_terminated)  # type: ignore[no-untyped-call]
